+++ parasite/scripts/production/core-creation/cegma.sh
@@ -6,8 +6,6 @@ if [ ! -f "$1" ] ; then echo "Usage: $0 <species/species.fa>" ; exit 1 ; fi
  
 fasta=$1
 species=$(basename $(dirname $fasta) )
-core_db=$($PARASITE_STAGING_MYSQL -e 'show databases' | grep "${species}_core_${PARASITE_VERSION}_${ENSEMBL_VERSION}" | head -n 1)
-if [ ! "core_db" ] ; then echo "Could not find core db for species $species " ; exit 1 ; fi
 
 tmp=$PARASITE_SCRATCH/cegma/WBPS${PARASITE_VERSION}/$species
 mkdir -pv $tmp
@@ -22,6 +20,8 @@ if ! [ "$cegma_complete" ] || ! [ "$cegma_partial" ] ; then
     echo "Could not parse: ./output.completeness_report"
     exit 1
 fi
+core_db=$($PARASITE_STAGING_MYSQL -e 'show databases' | grep "${species}_core_${PARASITE_VERSION}_${ENSEMBL_VERSION}" | head -n 1)
+if [ ! "core_db" ] ; then echo "Could not find core db for species $species " ; exit 1 ; fi
 ${PARASITE_STAGING_MYSQL}-ensrw $core_db -e 'delete from meta where meta_key like "assembly.cegma%"'
  
 ${PARASITE_STAGING_MYSQL}-ensrw $core_db -e "insert into meta (meta_key, meta_value) values (\"assembly.cegma_complete\", \"$cegma_complete\");"
